Refactoring conde
