+++ src/main/java/com/lealdev/restfull_spring_java/services/MathService.java
@@ -1,14 +1,17 @@
-package com.lealdev.restfull_spring_java;
+package com.lealdev.restfull_spring_java.services;
 
+import com.lealdev.restfull_spring_java.services.interfaces.IMathService;
 import com.lealdev.restfull_spring_java.exceptions.UnsuportedMathOperationException;
 import org.springframework.stereotype.Service;
+import static com.lealdev.restfull_spring_java.utils.MathUtils.convertToDouble;
+import static com.lealdev.restfull_spring_java.utils.MathUtils.isNumeric;
 
 @Service
-public class MathService implements IMathService{
+public class MathService implements IMathService {
 
     @Override
     public Double sum(String numberOne, String numberTwo) throws Exception {
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)){
+        if (!isNumeric(numberOne) || isNumeric(numberTwo)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return convertToDouble(numberOne) + convertToDouble(numberTwo);
@@ -16,7 +19,7 @@ public class MathService implements IMathService{
 
     @Override
     public Double subtraction(String numberOne, String numberTwo) throws Exception {
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)){
+        if (isNumeric(numberOne) || isNumeric(numberTwo)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return convertToDouble(numberOne) - convertToDouble(numberTwo);
@@ -24,7 +27,7 @@ public class MathService implements IMathService{
 
     @Override
     public Double average(String numberOne, String numberTwo) throws Exception {
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)){
+        if (isNumeric(numberOne) || isNumeric(numberTwo)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return (convertToDouble(numberOne) + convertToDouble(numberTwo)) / 2;
@@ -32,7 +35,7 @@ public class MathService implements IMathService{
 
     @Override
     public Double squareRoot(String number) throws Exception {
-        if (!isNumeric(number)){
+        if (isNumeric(number)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return Math.sqrt(convertToDouble(number));
@@ -40,7 +43,7 @@ public class MathService implements IMathService{
 
     @Override
     public Double multiplication(String numberOne, String numberTwo) throws Exception {
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)){
+        if (isNumeric(numberOne) || isNumeric(numberTwo)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return convertToDouble(numberOne) * convertToDouble(numberTwo);
@@ -48,22 +51,11 @@ public class MathService implements IMathService{
 
     @Override
     public Double division(String numberOne, String numberTwo) throws Exception {
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)){
+        if (isNumeric(numberOne) || isNumeric(numberTwo)){
             throw new UnsuportedMathOperationException("Please set a numeric value!");
         }
         return convertToDouble(numberOne) / convertToDouble(numberTwo);
     }
 
-    private Double convertToDouble(String strNumber) {
-        if (strNumber == null) return 0D;
-        String number = strNumber.replace(",",".");
-        if (isNumeric(number)) return Double.parseDouble(number);
-        return 0D;
-    }
-    private boolean isNumeric(String strNumber){
-        if (strNumber == null) return false;
-        String number = strNumber.replace(",",".");
-        return number.matches("[-+]?[0-9]*\\.?[0-9]+");
 
-    };
 }
